backend/pyproject.toml
[tool.poetry]
name = "truckbooking"
version = "0.1.0"
description = ""
authors = ["zeph <zephyrr2722@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.2.5"
psycopg2-binary = "^2.9.1"
python-decouple = "^3.4"
djangorestframework = "^3.12.4"
django-rest-knox = "^4.1.0"
gunicorn = "^20.1.0"
firebase-admin = "^5.0.2"
fcm-django = "^1.0.4"
django-cors-headers = "^3.8.0"
drf-yasg = "^1.20.0"
django-hosts = "^4.0"
Pillow = "^8.3.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
